requirements.txt
pydantic
uvicorn
fastapi
dotenv
langchain
langchain-huggingface
langchain-groq
langchain-openai
langchain-llm
langchain-ollama
langchain-aws
langchain-chroma
langchain-community
langchain-tools
groq
crewai
crewai-tools
crewai-adapters
crewai-tools[tools]
embedchain
crewai
pdfplumber
easyocr
sentence-transformers
faiss-cpu
numpy
pillow
python-multipart
streamlit